pyproject.toml
[project]
name = "agentic-resume-screener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.68.0",
    "ipykernel>=7.2.0",
    "langchain-core>=1.2.20",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.3",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2",
    "pytest>=9.0.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
]
